requirements.txt
requests
feedparser
beautifulsoup4
python-dotenv
lxml
